공유 uri ,계획정보

+++ michiki/src/main/java/com/michiki/michiki/place/dto/PlaceRecommendationRequestDto.java
@@ -8,12 +8,12 @@ import lombok.Setter;
 @Setter
 
 public class PlaceRecommendationRequestDto {
-    @NotNull
-    private Double centerLatitude;
+        @NotNull
+        private Double centerLatitude;
 
-    @NotNull
-    private Double centerLongitude;
+        @NotNull
+        private Double centerLongitude;
 
-    @NotNull
-    private Float zoomLevel;
+        @NotNull
+        private Integer zoomLevel;
 }
